moved reducers into new store folder
