+++ src/store/reducers/dashboard.js
@@ -8,8 +8,8 @@ import {
   UPDATE_EVENT_DURATION,
   TOGGLE_BOOKING_MODAL,
   SET_IS_BEING_UPDATED,
-} from '../actionTypes';
-import eventsView from '../constants/eventsView';
+} from '../actions/actionTypes';
+import eventsView from '../../constants/eventsView';
 import moment from 'moment';
 
 const initialState = {
